MatrixPilotQuad_MarkWhitehorn branch: project MPQ6000dual, fixed unintentional type conversion
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPillotQuad_MarkWhitehorn@1920 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilotQuad/MPQ6000dual.X/main.c
@@ -243,7 +243,7 @@ void run_background_task()
 {
 
     // do stuff which doesn't belong in ISRs
-    static int lastUptime = 0;
+    static unsigned long lastUptime = 0;
     if ((uptime - lastUptime) >= HEARTBEAT_HZ / 20)
     { // at 20 Hz
         lastUptime = uptime;
